GCI-1435: Rationalise item options class hierarchy.
